Add german dictionary
This adds the German dictionary and the top 1000 common German words. Umlauts are written as ae, oe, ue, ss.

I have added a directory for word lists and adjusted the imports, so nothing should change, as the default is still English.

I didn't have time to add a language switcher yet, but I'm going to bed and I thought someone else may benefit from the dictionary.

To use it locally, just change the three imports (two in Game.tsx and one in util.ts) to include the German dictionary instead.

The dictionary is based off of the OpenOffice dictionary and is licensed under the GPLv2: https://extensions.openoffice.org/en/project/german-de-de-frami-dictionaries