Ultra Hard: enforce letter counts even without absent clues
For example, with a target word of RALLY ([challenge link]), guessing
LEVEL gives [elsewhere, absent, absent, absent, elsewhere].
This commit makes guessing WORLD afterwards a violation, as there must
be at least two Ls in the target word.

This commit also makes it clearer that, when there is an absent clue,
the count of the letter must be exactly (number of non-absent clues with
that letter).

[challenge link]: https://hellowordl.net/?challenge=cmFsbHk